codensed tests
